Documentation

LyonCalcul.jl

Documentation for LyonCalcul.jl

LyonCalcul.Advection — Type.

Advection(n, p, delta)

n :: Number of points.
p :: Spline degree.
delta :: space size step.

source

LyonCalcul.Advection — Method.

advection! = Advection( n, p, delta)
advection!( f, alpha )

Create a function to compute the interpolating spline 
of degree p of odd
degree of a 1D function f on a periodic uniform mesh, at
all points x after a displacement alpha. 
Input f type is Vector{Float64} and is updated inplace.

source

LyonCalcul.bspline — Method.

bspline(p, j, x)

Return the value at x in [0,1[ of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

source